Use UCS-4 for rendering and display 6-digit replacement characters.
As we introduce support for discovering and rendering astral
characters, ensure that we pass UCS-4 to the relevant Font Manager
APIs and extend our replacement hex code generation to emit
6 digits for codepoints outside the Basic Multilingual Plane.

This has necessitated a change to the API of the callback function
provided to rufl_paint_callback(). Where, previously, a 16 bit
UCS-2 string was exposed, we now expose UCS-4.

+++ include/rufl.h
@@ -124,7 +124,7 @@ rufl_code rufl_split(const char *font_family, rufl_style font_style,
 /** Type of callback function for rufl_paint_callback(). */
 typedef void (*rufl_callback_t)(void *context,
 		const char *font_name, unsigned int font_size,
-		const char *s8, unsigned short *s16, unsigned int n,
+		const char *s8, unsigned int *s32, unsigned int n,
 		int x, int y);
